Move data fetching logic to custom hook
