+++ components/crewInformation/CrewInformationContainer.tsx
@@ -1,13 +1,12 @@
 'use client';
 
 import React, { useMemo } from 'react';
-import { useQuery } from '@tanstack/react-query';
-import { CrewResponse, CrewResponseError } from '@/types/crewResponse';
 import { useParams } from 'next/navigation';
 import { CrewMembers } from './CrewMembers';
 import { CrewInformationLoading } from './CrewInformationLoading';
 import { Pagination } from '@/components/pagination/Pagination';
 import { CrewInformationError } from '../errors/CrewInformationError';
+import { useGetCrew } from '@/hooks/useGetCrew';
 
 export const CrewInformationContainer = () => {
 	const params = useParams();
@@ -15,26 +14,7 @@ export const CrewInformationContainer = () => {
 		return params && params.page ? Number(params.page) : 1;
 	}, [params]);
 
-	const {
-		data: crewInfomration,
-		isLoading,
-		error,
-	} = useQuery({
-		queryFn: async () => {
-			const res = await fetch(`/api/crew?page=${pageNumber}`);
-
-			if (!res.ok) {
-				const error = (await res.json()) as CrewResponseError;
-				throw new Error(error.message);
-			}
-
-			const data = (await res.json()) as CrewResponse;
-
-			return data;
-		},
-		enabled: !!params,
-		queryKey: ['geCrewInfomration', pageNumber],
-	});
+	const { data: crewInfomration, isLoading, error } = useGetCrew(pageNumber);
 
 	if (error) return <CrewInformationError error={error} />;
 	return (
