+++ tests/components/crewInformation/CrewInformationLoading.test.tsx
@@ -1,5 +1,5 @@
-import { CrewInformationLoading } from '@/components/crewInformation/CrewInformationLoading';
 import '@testing-library/jest-dom';
+import { CrewInformationLoading } from '@/components/crewInformation/CrewInformationLoading';
 import { render, screen } from '@testing-library/react';
 
 describe('CrewInformationLoading Component', () => {
